requirements.txt
lxml==3.4.1
PyYAML==3.11
requests==2.5.1
